Extract TapTempo.js
